# api/src/swagger/swagger.json
{
	"components": {
...
		"schemas": {
...
			"IToken": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"token": {
						"type": "string"
					},
					"createdToken": {
						"type": "number",
						"format": "double"
					},
					"expieredToken": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"token"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IPlatform": {
				"properties": {
					"idPlatformOfAdmin": {
						"type": "string"
					},
					"platformName": {
						"type": "string"
					},
					"platformId": {
						"type": "string"
					}
				},
				"required": [
					"idPlatformOfAdmin",
					"platformName",
					"platformId"
				],
				"type": "object"
			},
			"IAppProfile": {
				"properties": {
					"appProfileName": {
						"type": "string"
					},
					"appProfileBosConfigId": {
						"type": "string"
					},
					"appProfileAdminId": {
						"type": "string"
					}
				},
				"required": [
					"appProfileName",
					"appProfileBosConfigId",
					"appProfileAdminId"
				],
				"type": "object"
			},
			"IAppPlatformProfile": {
				"allOf": [
					{
						"$ref": "#/components/schemas/IPlatform"
					},
					{
						"properties": {
							"appProfile": {
								"$ref": "#/components/schemas/IAppProfile"
							}
						},
						"required": [
							"appProfile"
						],
						"type": "object"
					}
				]
			},
...
			"IUserProfile": {
				"properties": {
					"userProfileName": {
						"type": "string"
					},
					"userProfileBosConfigId": {
						"type": "string"
					},
					"userProfileAdminId": {
						"type": "string"
					}
				},
				"required": [
					"userProfileName",
					"userProfileBosConfigId",
					"userProfileAdminId"
				],
				"type": "object"
			},
			"IUserPlatformProfile": {
				"allOf": [
					{
						"$ref": "#/components/schemas/IPlatform"
					},
					{
						"properties": {
							"userProfile": {
								"$ref": "#/components/schemas/IUserProfile"
							}
						},
						"required": [
							"userProfile"
						],
						"type": "object"
					}
				]
			},
...
			"ICodeToken": {
				"allOf": [
					{
						"$ref": "#/components/schemas/IToken"
					},
					{
						"properties": {
							"platformList": {
								"items": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IAppPlatformProfile"
										},
										{
											"$ref": "#/components/schemas/IUserPlatformProfile"
										}
									]
								},
								"type": "array"
							},
							"applicationProfileList": {
								"items": {
									"type": "string"
								},
								"type": "array"
							},
							"userId": {
								"type": "string"
							},
							"applicationId": {
								"type": "string"
							}
						},
						"type": "object"
					}
				]
			},
			"ICodeBase": {
				"properties": {
					"usedAt": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "number",
						"format": "double"
					},
					"used": {
						"type": "boolean"
					},
					"code": {
						"type": "string"
					}
				},
				"required": [
					"usedAt",
					"createdAt",
					"used",
					"code"
				],
				"type": "object"
			},
			"ICodeResponse": {
				"allOf": [
					{
						"$ref": "#/components/schemas/ICodeBase"
					},
					{
						"properties": {
							"profiles": {
								"items": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IAppPlatformProfile"
										},
										{
											"$ref": "#/components/schemas/IUserPlatformProfile"
										}
									]
								},
								"type": "array"
							},
							"id": {
								"type": "string"
							}
						},
						"required": [
							"profiles",
							"id"
						],
						"type": "object"
					}
				]
			},
			"IUserProfileBase": {
				"properties": {
					"userProfileId": {
						"type": "string"
					},
					"platformId": {
						"type": "string"
					}
				},
				"required": [
					"userProfileId",
					"platformId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IAppProfileBase": {
				"properties": {
					"appProfileId": {
						"type": "string"
					},
					"platformId": {
						"type": "string"
					}
				},
				"required": [
					"appProfileId",
					"platformId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IProfile": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/IUserProfileBase"
					},
					{
						"$ref": "#/components/schemas/IAppProfileBase"
					}
				]
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "Spinal Organ Auth API",
		"version": "1.0.0",
		"description": "API for managing authentication and authorization in the Spinal Organ system",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Spinalcom",
			"email": "developers@spinalcom.com"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/codes/consume/{code}": {
			"post": {
				"operationId": "ConsumeCode",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ICodeToken"
										},
										{
											"properties": {
												"error": {
													"type": "string"
												}
											},
											"required": [
												"error"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Codes"
				],
				"security": [
					{
						"all": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "code",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/codes": {
			"get": {
				"operationId": "GetAllCode",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"items": {
												"$ref": "#/components/schemas/ICodeResponse"
											},
											"type": "array"
										},
										{
											"properties": {
												"error": {
													"type": "string"
												}
											},
											"required": [
												"error"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Codes"
				],
				"security": [
					{
						"jwt": [
							"authAdmin"
						]
					}
				],
				"parameters": []
			}
		},
		"/codes/create": {
			"post": {
				"operationId": "GenerateCode",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"items": {
												"$ref": "#/components/schemas/ICodeResponse"
											},
											"type": "array"
										},
										{
											"properties": {
												"error": {
													"type": "string"
												}
											},
											"required": [
												"error"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Codes"
				],
				"security": [
					{
						"jwt": [
							"authAdmin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"count": {
										"type": "number",
										"format": "double"
									},
									"profiles": {
										"anyOf": [
											{
												"$ref": "#/components/schemas/IProfile"
											},
											{
												"items": {
													"$ref": "#/components/schemas/IProfile"
												},
												"type": "array"
											}
										]
									}
								},
								"required": [
									"profiles"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/codes/getcode/{code}": {
			"get": {
				"operationId": "GetCode",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ICodeResponse"
										},
										{
											"properties": {
												"error": {
													"type": "string"
												}
											},
											"required": [
												"error"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Codes"
				],
				"security": [
					{
						"jwt": [
							"authAdmin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "code",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/codes/delete/{code}": {
			"delete": {
				"operationId": "RemoveCode",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ICodeResponse"
										},
										{
											"properties": {
												"error": {
													"type": "string"
												}
											},
											"required": [
												"error"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Codes"
				],
				"security": [
					{
						"jwt": [
							"authAdmin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "code",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/codes/delete": {
			"delete": {
				"operationId": "RemoveCodes",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"items": {
												"$ref": "#/components/schemas/ICodeResponse"
											},
											"type": "array"
										},
										{
											"properties": {
												"error": {
													"type": "string"
												}
											},
											"required": [
												"error"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Codes"
				],
				"security": [
					{
						"jwt": [
							"authAdmin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"codes": {
										"items": {
											"type": "string"
										},
										"type": "array"
									}
								},
								"required": [
									"codes"
								],
								"type": "object"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}